Add a initial orchestration engine.
The goal of this layer is to be able to provide a high
level api reception point for which requests for resources
(initially compute resources) will channel through. It will
handle the complex set of transitions (and unraveling) needed
on errors and state transitions so that the components provisioning
and consuming those resources do not have to be complicated by
such knowledge.

This layer will initially be targeted at working this model
into the compute 'create' call where it can be used to show
others of how such a model is better than what exists currently.

Blueprint instance-create-state-machine

See also: http://wiki.openstack.org/RepairingOwnership
See also: http://wiki.openstack.org/SchedulingSplitout (a dependent)